+++ src/main/php/text/ical/Event.class.php
@@ -7,7 +7,12 @@ class Event implements Object {
   use Event\is\Value;
   use Event\with\Builder;
 
-  private $organizer, $attendee, $description, $summary, $comment, $dtstart, $dtend, $uid, $class, $location, $status, $alarm;
+  private $organizer, $attendee;
+  private $description, $summary, $comment;
+  private $dtstart, $dtend, $dtstamp;
+  private $uid, $class, $priority, $transp, $sequence, $status;
+  private $location;
+  private $alarm;
 
   /**
    * Write this object
@@ -25,10 +30,14 @@ class Event implements Object {
       'summary'     => $this->summary,
       'dtstart'     => $this->dtstart,
       'dtend'       => $this->dtend,
+      'dtstamp'     => $this->dtstamp,
       'uid'         => $this->uid,
       'class'       => $this->class,
+      'priority'    => $this->priority,
+      'transp'      => $this->transp,
       'location'    => $this->location,
       'status'      => $this->status,
+      'sequence'    => $this->sequence,
       'alarm'       => $this->alarm
     ]);
   }